Add confirm account form to fosuser confirm account flow
The default fosuser behavior is to send a confirm token to the email address. When the link is followed, the account is immediately confirmed and the temporary token deleted. This is a problem because some email clients (yahoo) prefetch the link and the user account is confirmed before the user actually confirms it. This commit adds a form to the fosuser confirmAction controller so that the user has to click a button the website after they follow the emailed link.